Do not depend on gmt
If mkgmap is installed, gmt (which is not free software as far as I can
see) is only used to set the FID in a TYP file once. This can also be
achieved without gmt (in a somewhat hackish way).

If gmt is not found, fall back to this way, making gmt completely
optional.